requirements.txt
Flask
Flask-SocketIO
mysql-connector-python
psutil
scikit-learn
numpy
eventlet
gunicorn
reportlab